backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
python-dotenv==1.0.1
pyjwt==2.9.0
pycryptodome==3.21.0
httpx==0.27.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
